# openapi.yaml
openapi: 3.0.3

info:
  title: Pollo Asado API
  description: API REST para gestionar productos, clientes y pedidos de un restaurante de pollo asado.
  version: 1.0.0

servers:
  - url: http://localhost:3000
    description: Servidor local

tags:
  - name: Products
    description: Operaciones relacionadas con productos
  - name: Customers
    description: Operaciones relacionadas con clientes
  - name: Orders
    description: Operaciones relacionadas con pedidos

paths:

  /api/products:
    get:
      tags:
        - Products
      summary: Obtener todos los productos
      responses:
        '200':
          description: Lista de productos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'

    post:
      tags:
        - Products
      summary: Crear un producto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductInput'
      responses:
        '201':
          description: Producto creado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Datos requeridos faltantes

  /api/products/{id}:
    get:
      tags:
        - Products
      summary: Obtener un producto por ID
      parameters:
        - $ref: '#/components/parameters/IdParameter'
      responses:
        '200':
          description: Producto encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Producto no encontrado

    put:
      tags:
        - Products
      summary: Actualizar un producto
      parameters:
        - $ref: '#/components/parameters/IdParameter'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductInput'
      responses:
        '200':
          description: Producto actualizado correctamente
        '400':
          description: Datos requeridos faltantes
        '404':
          description: Producto no encontrado

    delete:
      tags:
        - Products
      summary: Eliminar un producto
      parameters:
        - $ref: '#/components/parameters/IdParameter'
      responses:
        '200':
          description: Producto eliminado correctamente
        '404':
          description: Producto no encontrado

  /api/customers:
    get:
      tags:
        - Customers
      summary: Obtener todos los clientes
      responses:
        '200':
          description: Lista de clientes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Customer'

    post:
      tags:
        - Customers
      summary: Crear un cliente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerInput'
      responses:
        '201':
          description: Cliente creado correctamente
        '400':
          description: Datos requeridos faltantes

  /api/customers/{id}:
    get:
      tags:
        - Customers
      summary: Obtener un cliente por ID
      parameters:
        - $ref: '#/components/parameters/IdParameter'
      responses:
        '200':
          description: Cliente encontrado
        '404':
          description: Cliente no encontrado

    put:
      tags:
        - Customers
      summary: Actualizar un cliente
      parameters:
        - $ref: '#/components/parameters/IdParameter'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerInput'
      responses:
        '200':
          description: Cliente actualizado correctamente
        '400':
          description: Datos requeridos faltantes
        '404':
          description: Cliente no encontrado

    delete:
      tags:
        - Customers
      summary: Eliminar un cliente
      parameters:
        - $ref: '#/components/parameters/IdParameter'
      responses:
        '200':
          description: Cliente eliminado correctamente
        '404':
          description: Cliente no encontrado

  /api/orders:
    get:
      tags:
        - Orders
      summary: Obtener todos los pedidos
      responses:
        '200':
          description: Lista de pedidos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'

    post:
      tags:
        - Orders
      summary: Crear un pedido
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderInput'
      responses:
        '201':
          description: Pedido creado correctamente
        '400':
          description: Datos requeridos faltantes

  /api/orders/{id}:
    get:
      tags:
        - Orders
      summary: Obtener un pedido por ID
      parameters:
        - $ref: '#/components/parameters/IdParameter'
      responses:
        '200':
          description: Pedido encontrado
        '404':
          description: Pedido no encontrado

    put:
      tags:
        - Orders
      summary: Actualizar un pedido
      parameters:
        - $ref: '#/components/parameters/IdParameter'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderInput'
      responses:
        '200':
          description: Pedido actualizado correctamente
        '400':
          description: Datos requeridos faltantes
        '404':
          description: Pedido no encontrado

    delete:
      tags:
        - Orders
      summary: Eliminar un pedido
      parameters:
        - $ref: '#/components/parameters/IdParameter'
      responses:
        '200':
          description: Pedido eliminado correctamente
        '404':
          description: Pedido no encontrado

components:

  parameters:
    IdParameter:
      name: id
      in: path
      required: true
      description: Identificador del recurso
      schema:
        type: integer
        example: 1

  schemas:

    Product:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Pollo Asado Entero
        price:
          type: number
          example: 28000

    ProductInput:
      type: object
      required:
        - name
        - price
      properties:
        name:
          type: string
          example: Pollo Asado Entero
        price:
          type: number
          example: 28000

    Customer:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Natalia Omaña
        email:
          type: string
          format: email
          example: natalia@example.com
        phone:
          type: string
          example: "3001234567"

    CustomerInput:
      type: object
      required:
        - name
        - email
        - phone
      properties:
        name:
          type: string
          example: Natalia Omaña
        email:
          type: string
          format: email
          example: natalia@example.com
        phone:
          type: string
          example: "3001234567"

    Order:
      type: object
      properties:
        id:
          type: integer
          example: 1
        customerId:
          type: integer
          example: 1
        products:
          type: array
          items:
            type: object
            properties:
              productId:
                type: integer
                example: 1
              quantity:
                type: integer
                example: 2
        total:
          type: number
          example: 56000
        status:
          type: string
          example: pending

    OrderInput:
      type: object
      required:
        - customerId
        - products
        - total
        - status
      properties:
        customerId:
          type: integer
          example: 1
        products:
          type: array
          items:
            type: object
            required:
              - productId
              - quantity
            properties:
              productId:
                type: integer
                example: 1
              quantity:
                type: integer
                example: 2
        total:
          type: number
          example: 56000
        status:
          type: string
          example: pending